reduced duplications and added missing booking infosectionfile
