+++ apps/native/src/features/booking-orders/components/shared/InfoSection.tsx
@@ -17,45 +17,8 @@ interface Props {
 function InfoRow({ icon: Icon, label, onPress, value }: InfoSectionRow) {
   const themeColors = useThemeColors();
   const color = onPress ? themeColors.primary : themeColors.textPrimary;
-
-  if (onPress) {
-    return (
-      <TouchableOpacity className="flex-row items-center gap-3" activeOpacity={0.7} onPress={onPress}>
-        <View
-          className="h-10 w-10 items-center justify-center rounded-xl"
-          style={{ backgroundColor: `${themeColors.primary}12` }}
-        >
-          <Icon size={18} color={themeColors.primary} strokeWidth={2} />
-        </View>
-        <View className="flex-1">
-          <Text
-            style={{
-              fontSize: 11,
-              color: themeColors.textMuted,
-              textTransform: "uppercase",
-              letterSpacing: 0.5,
-            }}
-          >
-            {label}
-          </Text>
-          <Text
-            style={{
-              fontFamily: "GoogleSans_600SemiBold",
-              fontSize: 14,
-              color,
-              marginTop: 1,
-            }}
-            numberOfLines={2}
-          >
-            {value}
-          </Text>
-        </View>
-      </TouchableOpacity>
-    );
-  }
-
-  return (
-    <View className="flex-row items-center gap-3">
+  const content = (
+    <>
       <View
         className="h-10 w-10 items-center justify-center rounded-xl"
         style={{ backgroundColor: `${themeColors.primary}12` }}
@@ -85,8 +48,22 @@ function InfoRow({ icon: Icon, label, onPress, value }: InfoSectionRow) {
           {value}
         </Text>
       </View>
-    </View>
+    </>
   );
+
+  if (onPress) {
+    return (
+      <TouchableOpacity
+        className="flex-row items-center gap-3"
+        activeOpacity={0.7}
+        onPress={onPress}
+      >
+        {content}
+      </TouchableOpacity>
+    );
+  }
+
+  return <View className="flex-row items-center gap-3">{content}</View>;
 }
 
 export default function InfoSection({ rows }: Props) {
